## What does this PR do ?

Adjust resource limits and decrease the number of chunk iterations in the `tests/streaming/test_dataset.py::test_dataset_for_text_tokens_with_large_num_chunks` test to improve test execution time.


Partial work on https://github.com/Lightning-AI/litData/issues/612

**Before**
```python
# ubuntu
195.33s call     tests/streaming/test_dataset.py::test_dataset_for_text_tokens_with_large_num_chunks

#macos
345.54s call     tests/streaming/test_dataset.py::test_dataset_for_text_tokens_with_large_num_chunks
```

**After**
```python
# ubuntu
31.33s call     tests/streaming/test_dataset.py::test_dataset_for_text_tokens_with_large_num_chunks

#macos
56.55s call     tests/streaming/test_dataset.py::test_dataset_for_text_tokens_with_large_num_chunks
```